Extract FallbackRequestResolver
