+++ DNS/Server/DnsServer.cs
@@ -163,24 +163,5 @@ namespace DNS.Server {
                 finally { OnErrored(e); }
             }
         }
-
-        private class FallbackRequestResolver : IRequestResolver {
-            private IRequestResolver[] resolvers;
-
-            public FallbackRequestResolver(params IRequestResolver[] resolvers) {
-                this.resolvers = resolvers;
-            }
-
-            public async Task<IResponse> Resolve(IRequest request) {
-                IResponse response = null;
-
-                foreach (IRequestResolver resolver in resolvers) {
-                    response = await resolver.Resolve(request);
-                    if (response.AnswerRecords.Count > 0) break;
-                }
-
-                return response;
-            }
-        }
     }
 }
